% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combinations}
\alias{combinations}
\title{Combinations of input arguments}
\usage{
combinations(...)
}
\arguments{
\item{...}{Any number of atomic vectors.}
}
\value{
A data frame with columns named as the inputs, containing all input
combinations.
}
\description{
Combinations of input arguments
}
\examples{
combinations(a = 1:2, b = letters[1:3], c = LETTERS[1:2])
}
