% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constDiagMatrix}
\alias{constDiagMatrix}
\title{Constant plus diagonal matrix}
\usage{
constDiagMatrix(dim = 1, diag = 1, offDiag = 0)
}
\arguments{
\item{dim}{Dimension}

\item{diag}{Value at the diagonal}

\item{offDiag}{Value at off-diagonals}
}
\value{
A \code{dim}-by-\code{dim} matrix
}
\description{
Specify a matrix with constant values on the diagonal and on the off-diagonals.
Such matrices can be used to vary the degree of dependency in covariate matrices,
for example when evaluating accuracy of KL-divergence estimation algorithms.
}
\examples{
constDiagMatrix(dim = 3, diag = 1, offDiag = 0.9)

}
