% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_two_sample}
\alias{is_two_sample}
\title{Detect if a one- or two-sample problem is specified}
\usage{
is_two_sample(Y, q)
}
\arguments{
\item{Y}{A vector, matrix, data frame or \code{NULL}}

\item{q}{A function or \code{NULL}.}
}
\value{
\code{TRUE} for a two-sample problem (i.e., \code{Y} non-null and \code{q = NULL})
and \code{FALSE} for a one-sample problem (i.e., \code{Y = NULL} and \code{q} non-null).
}
\description{
Detect if a one- or two-sample problem is specified
}
