% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld-analytical.R
\name{kl_divergence_gaussian}
\alias{kl_divergence_gaussian}
\title{Analytical KL divergence for two uni- or multivariate Gaussian distributions}
\usage{
kl_divergence_gaussian(mu1, sigma1, mu2, sigma2)
}
\arguments{
\item{mu1}{A numeric vector (mean of true Gaussian)}

\item{sigma1}{A s.p.d. matrix (Covariance matrix of true Gaussian)}

\item{mu2}{A numeric vector (mean of approximate Gaussian)}

\item{sigma2}{A s.p.d. matrix  (Covariance matrix of approximate Gaussian)}
}
\value{
A scalar (the Kullback-Leibler divergence)
}
\description{
This function computes $D_{KL}(p||q)$, where $p\sim \mathcal{N}(\mu_1,\Sigma_1)$
and $q\sim \mathcal{N}(\mu_2,\Sigma_2)$.
}
