% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld-uncertainty.R
\name{kld_ci_subsampling}
\alias{kld_ci_subsampling}
\title{Uncertainty of KL divergence estimate using Politis/Romano's subsampling bootstrap.}
\usage{
kld_ci_subsampling(
  X,
  Y = NULL,
  q = NULL,
  estimator = kld_est_nn,
  B = 500L,
  alpha = 0.05,
  subsample.size = function(x) x^(2/3),
  convergence.rate = sqrt
)
}
\arguments{
\item{X, Y}{\code{n}-by-\code{d} and \code{m}-by-\code{d} matrices, representing \code{n} samples from
the true distribution \eqn{P} and \code{m} samples from the approximate distribution
\eqn{Q}, both in \code{d} dimensions. Vector input is treated as a column matrix.
\code{Y} can be left blank if \code{q} is specified (see below).}

\item{q}{The density function of the approximate distribution \eqn{Q}. Either
\code{Y} or \code{q} must be specified.}

\item{estimator}{The Kullback-Leibler divergence estimation method; a
function expecting two inputs (\code{X} and \code{Y} or \code{q}, depending on arguments
provided). Defaults to \code{kld_est_nn}.}

\item{B}{Number of bootstrap replicates (default: \code{500}), the larger, the
more accurate, but also more computationally expensive.}

\item{alpha}{Error level, defaults to \code{0.05}.}

\item{subsample.size}{A function specifying the size of the subsamples,
defaults to \eqn{f(x) = x^{2/3}}.}

\item{convergence.rate}{A function computing the convergence rate of the
estimator as a function of sample sizes. Defaults to \eqn{f(x) = x^{1/2}}.}
}
\value{
A list with the fields \code{"est"} (the estimated KL divergence),
\code{"boot"} (a length \code{B} numeric vector with KL divergence estimates on
the bootstrap subsamples), and \code{"ci"} (a length \code{2} vector containing the
lower and upper limits of the estimated confidence interval).
}
\description{
This function computes a confidence interval for KL divergence based on the
subsampling bootstrap by Politis and Romano. The calculated interval has
asymptotic coverage \eqn{1 - \alpha} as long as \eqn{b_n/n\rightarrow 0},
\eqn{b_n\rightarrow\infty} and \eqn{\frac{\tau_b}{\tau_n}\rightarrow 0}.
}
\details{
Reference:

Politis and Romano, "Large sample confidence regions based on subsamples under
minimal assumptions", The Annals of Statistics, Vol. 22, No. 4 (1994).
}
\examples{
# 1D Gaussian (one- and two-sample problems)
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
q <- function(x) dnorm(x, mean =1, sd = 2)
kld_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
kld_est_nn(X, Y = Y)
kld_est_nn(X, q = q)
kld_ci_subsampling(X, Y)$ci
kld_ci_subsampling(X, q = q)$ci

}
