% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld-uncertainty.R
\name{kld_ci_subsampling2}
\alias{kld_ci_subsampling2}
\title{Parallel version of kld_ci_subsampling, for testing (unfinished, not exported)}
\usage{
kld_ci_subsampling2(
  X,
  Y = NULL,
  q = NULL,
  estimator = kld_est_nn,
  B = 500L,
  alpha = 0.05,
  subsample.size = function(x) x^(2/3),
  convergence.rate = sqrt,
  method = c("quantile", "se"),
  n.cores = 1L
)
}
\description{
Parallel version of kld_ci_subsampling, for testing (unfinished, not exported)
}
