% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld-estimation-interfaces.R
\name{kld_est}
\alias{kld_est}
\title{Kullback-Leibler divergence estimator for discrete, continuous or mixed data.}
\usage{
kld_est(X, Y, ..., vartype = NULL)
}
\arguments{
\item{X, Y}{Data frames or matrices with the same number of columns \code{d}
(multivariate samples), or numeric/character vectors (univariate samples,
i.e. \code{d=1}).}

\item{...}{Inputs passed on to function \code{\link[=kld_cont]{kld_cont()}}.}

\item{vartype}{A length \code{d} character vector, with \code{vartype[i] = "c"} meaning
the \code{i}-th variable is continuous, and \code{vartype[i] = "d"} meaning it is
discrete. If unspecified, \code{vartype} is \code{"c"} for numeric columns and \code{"d"}
for character or factor columns. This default will not work if levels of
discrete variables are encoded using numbers (e.g., \code{0} for females and
\code{1} for males) or for count data.}
}
\value{
A scalar, the estimated Kullback-Leibler divergence \eqn{D_{KL}(P||Q)}.
}
\description{
For two mixed continuous/discrete distributions with densities \eqn{p} and
\eqn{q}, and denoting \eqn{x = (x_\text{c},x_\text{d})}, the Kullback-Leibler
divergence \eqn{D_{KL}(p||q)} is given as
\deqn{D_{KL}(p||q) = \sum_{x_d} \int p(x_c,x_d) \log\left(\frac{p(x_c,x_d)}{q(x_c,x_d)}\right)dx_c.}
Conditioning on the discrete variables \eqn{x_d}, this can be re-written as
\deqn{D_{KL}(p||q) = \sum_{x_d} p(x_d) D_{KL}\big(p(\cdot|x_d)||q(\cdot|x_d)\big) +
D_{KL}\big(p_{x_d}||q_{x_d}\big).}
Here, the terms
\deqn{D_{KL}\big(p(\cdot|x_d)||q(\cdot|x_d)\big)}
are approximated via nearest neighbour- or kernel-based density estimates on
the datasets \code{X} and \code{Y} stratified by the discrete variables, and
\deqn{D_{KL}\big(p_{x_d}||q_{x_d}\big)}
is approximated using relative frequencies.
}
\examples{
# 2D example
X <- data.frame(cont  = rnorm(10),
                discr = c(rep('a',4),rep('b',6)))
Y <- data.frame(cont  = c(rnorm(5), rnorm(5, sd = 2)),
                discr = c(rep('a',5),rep('b',5)))
kld_est(X,Y)
}
