% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld-estimation-continuous.R
\name{kld_est_densratio}
\alias{kld_est_densratio}
\title{Density-ratio-based estimation of Kullback-Leibler divergence}
\usage{
kld_est_densratio(X, Y, ...)
}
\arguments{
\item{X, Y}{\code{n}-by-\code{d} and \code{m}-by-\code{d} matrices, representing \code{n} samples from
the true distribution \eqn{P} and \code{m} samples from the approximate distribution
\eqn{Q}, both in \code{d} dimensions. Vector input is treated as a column matrix.}

\item{...}{Additional options to be passed to function \code{densratio}.}
}
\value{
A scalar, the estimated Kullback-Leibler divergence \eqn{D_{KL}(P||Q)}.
}
\description{
This estimation method approximates the ratio of densities of the unknown
distributions \eqn{P} and \eqn{Q}  using package 'densratio'.
}
\examples{
# KL-D between two samples from 1D Gaussians:
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
kl_div_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
kldest_densratio(X,Y)
kldest_densratio(X,Y, method = "KLIEP")
kldest_densratio(X,Y, method = "RuLSIF")
}
