% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld-estimation-discrete.R
\name{kld_est_discrete}
\alias{kld_est_discrete}
\title{Plug-in KL divergence estimator for samples from discrete distributions}
\usage{
kld_est_discrete(X, Y = NULL, q = NULL)
}
\arguments{
\item{X, Y}{Two samples from discrete distributions, specified as vectors,
matrices or data frames. Argument \code{Y} can be omitted if argument \code{q} is
given (see below).}

\item{q}{The probability mass function of the approximate distribution \eqn{Q}.}
}
\value{
A scalar, the estimated Kullback-Leibler divergence \eqn{D_{KL}(P||Q)}.
}
\description{
Plug-in KL divergence estimator for samples from discrete distributions
}
\examples{
# 1D example
X <- c(rep('M',5),rep('F',5))
Y <- c(rep('M',6),rep('F',4))
kld_est_discrete(X, Y)
}
