% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld-estimation-continuous.R
\name{kld_est_gknn}
\alias{kld_est_gknn}
\title{Generalized nearest-neighbour KL divergence estimation from Wang et al. (2009)}
\usage{
kld_est_gknn(X, Y, max.k = 50, warn.max.k = TRUE)
}
\arguments{
\item{X, Y}{\code{n}-by-\code{d} and \code{m}-by-\code{d} matrices, representing \code{n} samples from
the true distribution \eqn{P} and \code{m} samples from the approximate distribution
\eqn{Q}, both in \code{d} dimensions. Vector input is treated as a column matrix.}

\item{max.k}{Maximum numbers of nearest neighbours to compute (default: \code{50}).
A larger \code{max.k} may yield a more accurate KL-D estimate (see \code{warn.max.k}),
but will always increase the computational cost.}

\item{warn.max.k}{If \code{TRUE} (the default), warn if \code{max.k} is such that more
than \code{max.k} neighbours are within the \code{eps} neighbourhood for some data
points. In this case, only the first \code{max.k} neighbours will be counted.
As a consequence, \code{max.k} may required to be increased.}
}
\value{
A scalar, the estimated Kullback-Leibler divergence \eqn{D_{KL}(P||Q)}.
}
\description{
This is the generalized k-NN based KL divergence estimator, eq. (29).
\code{TODO}: more details on the algorithm!
}
\details{
Reference:
Wang, Kulkarni and Verdú, "Divergence Estimation for Multidimensional
Densities Via k-Nearest-Neighbor Distances", IEEE Transactions on Information
Theory, Vol. 55, No. 5 (2009). DOI: https://doi.org/10.1109/TIT.2009.2016060
}
\examples{
# KL-D between two samples from 1-D Gaussians:
X <- rnorm(100)
Y <- rnorm(100, mean = 1, sd = 2)
kld_gaussian(mu1 = 0, sigma1 = 1, mu2 = 1, sigma2 = 2^2)
kld_est_kde1(X, Y)
kld_est_1nn(X, Y)
kld_gknn(X, Y)

# KL-D between two samples from 2-D Gaussians:
X1 <- rnorm(100)
X2 <- rnorm(100)
Y1 <- rnorm(100)
Y2 <- Y1 + rnorm(100)
X <- cbind(X1,X2)
Y <- cbind(Y1,Y2)
kld_gaussian(mu1 = rep(0,2), sigma1 = diag(2),
             mu2 = rep(0,2), sigma2 = matrix(c(1,1,1,2),nrow=2))
kld_est_1nn(X, Y)
kld_est_gknn(X, Y)
}
