% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld-estimation.R
\name{kldest_density_vectorized}
\alias{kldest_density_vectorized}
\title{Vectorized version of kldest_density, currently under development}
\usage{
kldest_density_vectorized(
  X,
  Y,
  hX = NULL,
  hY = NULL,
  rule = c("Silverman", "Scott")
)
}
\description{
Vectorized version of kldest_density, currently under development
}
