% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kldest-package.R
\docType{package}
\name{kldest-package}
\alias{kldest}
\alias{kldest-package}
\title{kldest: Sample-based estimation of Kullback-Leibler divergence}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Niklas Hartung \email{niklas.hartung@gmail.com} (\href{https://orcid.org/0000-0002-4000-6525}{ORCID})

}
\keyword{internal}
