% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kldest-package.R
\docType{package}
\name{kldest-package}
\alias{kldest}
\alias{kldest-package}
\title{kldest: Sample-based estimation of Kullback-Leibler divergence}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
