% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kldest-package.R
\docType{package}
\name{kldest-package}
\alias{kldest}
\alias{kldest-package}
\title{kldest: Sample-based estimation of Kullback-Leibler divergence}
\description{
A collection of estimation algorithms of Kullback-Leibler divergence, based on one or two samples, and including uncertainty quantification. Distributions can be uni- or multivariate and continous, discrete or mixed.
}
\author{
\strong{Maintainer}: Niklas Hartung \email{niklas.hartung@gmail.com} (\href{https://orcid.org/0000-0002-4000-6525}{ORCID})

}
\keyword{internal}
