% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kldest-package.R
\docType{package}
\name{kldest-package}
\alias{kldest}
\alias{kldest-package}
\title{kldest: Sample-Based Estimation of Kullback-Leibler Divergence}
\description{
Estimation algorithms for Kullback-Leibler divergence between two probability distributions, based on one or two samples, and including uncertainty quantification. Distributions can be uni- or multivariate and continuous, discrete or mixed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://niklhart.github.io/kldest/}
  \item Report bugs at \url{https://github.com/niklhart/kldest/issues}
}

}
\author{
\strong{Maintainer}: Niklas Hartung \email{niklas.hartung@gmail.com} (\href{https://orcid.org/0000-0002-4000-6525}{ORCID}) [copyright holder]

}
\keyword{internal}
